+++ mobile/android/android-components/components/feature/session-bundling/src/main/java/mozilla/components/feature/session/bundling/db/BundleDao.kt
@@ -26,11 +26,11 @@ internal interface BundleDao {
     @Delete
     fun deleteBundle(bundle: BundleEntity)
 
-    @Query("SELECT * FROM bundles ORDER BY saved_at DESC LIMIT :limit")
-    fun getBundles(limit: Int = 20): LiveData<List<BundleEntity>>
+    @Query("SELECT * FROM bundles WHERE saved_at >= :since ORDER BY saved_at DESC LIMIT :limit")
+    fun getBundles(since: Long, limit: Int): LiveData<List<BundleEntity>>
 
-    @Query("SELECT * FROM bundles")
-    fun getBundlesPaged(): DataSource.Factory<Int, BundleEntity>
+    @Query("SELECT * FROM bundles WHERE saved_at >= :since")
+    fun getBundlesPaged(since: Long): DataSource.Factory<Int, BundleEntity>
 
     @Query("SELECT * FROM bundles WHERE saved_at >= :since ORDER BY saved_at DESC LIMIT 1")
     fun getLastBundle(since: Long): BundleEntity?