+++ mobile/android/android-components/docs/changelog.md
@@ -57,6 +57,9 @@ permalink: /changelog/
   * 🆕 New component to be used for helpers used in instrumented (on device) tests (`src/androidTest`). This component complements `support-test` which is focused on helpers used in local unit tests (`src/test`).
   * Added helper `LiveData.awaitValue()` which subscribes to the `LiveData` object and blocks until a value was observed. Returns the value or throws an `InterruptedException` if no value was observed (customizable timeout).
 
+* **feature-session-bundling**
+  * Added optional `since` parameter to `SessionBundleStorage.bundles()` and `SessionBundleStorage.bundlesPaged()`.
+
 # 0.44.0
 
 * [Commits](https://github.com/mozilla-mobile/android-components/compare/v0.43.0...v0.44.0)